+++ .github/CHANGELOG.md
@@ -1,12 +1,23 @@
 # Current develop
 
 ### New features
-  
+* Added one new private CircuitComponent ``_DsMap`` to convert the representations. 
+  [(#385)](https://github.com/XanaduAI/MrMustard/pull/385)
+
+* Added the function `phase_space(s)` for `State` class to take out the s-parametrized phase space representations' variables covariance matrix and means vector of the state. 
+  [(#385)](https://github.com/XanaduAI/MrMustard/pull/385)
+
+* Added `sort` function to math backends.
+
 ### Breaking changes
 
 ### Improvements
 
 ### Bug fixes
+* Fix the bug in the order of indices of the triples for DsMap CircuitComponent. 
+  [(#385)](https://github.com/XanaduAI/MrMustard/pull/385)
+
+* Ensure all symplectic eigenvalues are returned by the `symplectic_eigenvals` function.
 
 ### Documentation
 
@@ -16,6 +27,7 @@
 [Samuele Ferracin](https://github.com/SamFerracin)
 [Yuan Yao](https://github.com/sylviemonet)
 [Filippo Miatto](https://github.com/ziofil)
+[Austin Lund](https://github.com/aplund)
 
 
 ---